requirements.txt
PyYAML>=6.0
trafilatura>=1.9.0
